moved CreateMrjobZipTestCase to test_bin.py
