+++ tests/test_bin.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Test the runner base class MRJobBinRunner"""
+import compileall
 import inspect
 import os
 import signal
@@ -1476,3 +1477,32 @@ class SparkSubmitArgsTestCase(SandboxedTestCase):
             self.assertRaises(
                 TypeError,
                 runner._spark_submit_args, 0)
+
+
+class CreateMrjobZipTestCase(SandboxedTestCase):
+
+    def test_create_mrjob_zip(self):
+        with no_handlers_for_logger('mrjob.runner'):
+            with LocalMRJobRunner(conf_paths=[]) as runner:
+                mrjob_zip_path = runner._create_mrjob_zip()
+                mrjob_zip = ZipFile(mrjob_zip_path)
+                contents = mrjob_zip.namelist()
+
+                for path in contents:
+                    self.assertEqual(path[:6], 'mrjob/')
+
+                self.assertIn('mrjob/job.py', contents)
+                for filename in contents:
+                    self.assertFalse(filename.endswith('.pyc'),
+                                     msg="%s ends with '.pyc'" % filename)
+
+    def test_mrjob_zip_compiles(self):
+        runner = LocalMRJobRunner()
+        with no_handlers_for_logger('mrjob.runner'):
+            mrjob_zip = runner._create_mrjob_zip()
+
+        ZipFile(mrjob_zip).extractall(self.tmp_dir)
+
+        self.assertTrue(
+            compileall.compile_dir(os.path.join(self.tmp_dir, 'mrjob'),
+                                   quiet=1))
